refactor: Move mappers
