+++ backend_rust/src/adapters.rs
@@ -1,1 +1,2 @@
-pub mod api;
+pub mod recipe_mapper;
+pub mod step_mapper;
