--- backend_rust/src/adapters/api.rs
@@ -1,2 +1,0 @@
-pub mod recipe_mapper;
-pub mod step_mapper;
