+++ backend_rust/src/adapters/step_mapper.rs
